Update error logging to use ILogger's settings
Logic was included to include the full HTTP response from the API if the
verbose flag was set, but there was no way to set this flag by default.
Now, the response handler will use the client's logger level to
determine what level of response to include. In this instance, full HTTP
responses are included in the log if the logger level is set to DEBUG.