Deserialize derived class data
This will attempt to deserialize results into the odata.type that Graph responds with instead of the class that the request builder provides. This allows us to respond with a derived object to a base object call. It also allows a developer to successfully cast the response object from the base type to the expected derived type.